+++ src/sr2silo/storage/__init__.py
@@ -1,0 +1,7 @@
+""""Implements storage specific actions, such as uploading for S3."""
+
+from __future__ import annotations
+
+from sr2silo.storage.s3_storage import upload_file_to_s3
+
+__all__ = ["upload_file_to_s3"]
