+++ tests/storage/test_s3_storage.py
@@ -6,7 +6,7 @@ import boto3
 import pytest
 from moto import mock_aws
 
-from sr2silo.s3 import compress_bz2, upload_file_to_s3
+from sr2silo.storage.s3_storage import upload_file_to_s3
 
 
 @pytest.fixture
@@ -18,28 +18,6 @@ def s3_client():
         # Create a mock bucket
         s3.create_bucket(Bucket="test-bucket")
         yield s3
-
-
-def test_compress_bz2(tmp_path):
-    """Test the compress_bz2 function."""
-    # Create a test file
-    test_file = tmp_path / "test.txt"
-    test_file.write_text(
-        """
-        This is a test file – it should compress well.
-        This is a test file – it should compress well.
-        This is a test file – it should compress well.
-        """
-    )
-
-    # Compress the test file
-    compressed_file = tmp_path / "test.txt.bz2"
-    compress_bz2(test_file, compressed_file)
-
-    # Check that the compressed file exists
-    assert compressed_file.exists()
-    # Check that the compressed file is smaller than the original file
-    assert compressed_file.stat().st_size < test_file.stat().st_size
 
 
 def test_upload_file_to_s3(tmp_path, s3_client):
